fix: merge conflict issues
