+++ apps/web/src/components/DocumentWithEvaluations/index.tsx
@@ -1,10 +1,6 @@
 "use client";
 
-import {
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { useEffect, useMemo, useState } from "react";
 
 import { useEvaluationRerun } from "@/shared/hooks/useEvaluationRerun";
 import {
@@ -15,10 +11,7 @@ import { HEADER_HEIGHT_PX } from "@/shared/utils/ui/constants";
 
 import { EvaluationView } from "./components";
 import { EmptyEvaluationsView } from "./components/EmptyEvaluationsView";
-import type {
-  DocumentWithReviewsProps,
-  EvaluationState,
-} from "./types";
+import type { DocumentWithReviewsProps, EvaluationState } from "./types";
 
 export function DocumentWithEvaluations({
   document,
@@ -102,10 +95,7 @@ export function DocumentWithEvaluations({
   }, []);
 
   return (
-    <div
-      className="flex h-full flex-col"
-      style={{ height: `calc(100vh - ${HEADER_HEIGHT_PX}px)` }}
-    >
+    <div className="flex h-full flex-col mt-4">
       {hasEvaluations && evaluationState ? (
         <EvaluationView
           evaluationState={evaluationState}
